<fix>(loader): fix windows loader using.

+++ src/main/java/org/fisco/bcos/sdk/jni/common/JniLibLoader.java
@@ -20,9 +20,11 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.lang.reflect.Field;
+import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 import java.util.Objects;
+import java.util.Scanner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -40,6 +42,8 @@ public final class JniLibLoader {
 
   public static final String NATIVE_RESOURCE_LIB_NAME = "bcos-sdk-jni";
   public static final String NATIVE_RESOURCE_HOME = "/META-INF/native";
+  public static final String NATIVE_WIN_DEPS_DIR = NATIVE_RESOURCE_HOME + "/win/";
+  public static final String WIN_DEPS_FILE_LIST = "file.list";
 
   public static final String OS_NAME = getOs();
   public static final String ARCH_NAME = getArch();
@@ -142,10 +146,10 @@ public final class JniLibLoader {
     if (!loadLibFromFsOk) {
       try {
         logger.info("try to load library from jar");
-        loadLibraryFromJar(NATIVE_RESOURCE_LIB_NAME);
         if (Objects.equals(getOs(), WIN)) {
           loadWinDepsLibraryFromJar();
         }
+        loadLibraryFromJar(NATIVE_RESOURCE_LIB_NAME);
       } catch (Exception e1) {
         logger.error("unable to load library from fs, e: ", e1);
         throw new UnsatisfiedLinkError(e1.toString());
@@ -226,42 +230,40 @@ public final class JniLibLoader {
   }
 
   public static void loadWinDepsLibraryFromJar()
-      throws IOException, NoSuchFieldException, IllegalAccessException {
+      throws IOException, NoSuchFieldException {
     File tempDir = new File(WORKDIR, String.valueOf(System.nanoTime()));
     tempDir.mkdirs();
 
-    String winResource = NATIVE_RESOURCE_HOME + "/win/";
-    File winDir = new File(JniLibLoader.class.getResource(winResource).getFile());
-    // get all lib file in /native/win/
-    for (File lib : winDir.listFiles()) {
-      String libName = lib.getName();
-      File tempFile = new File(tempDir, libName);
-      logger.info("tempDir: {}, tempFile: {}", tempDir, tempFile);
-      try {
-        Files.copy(lib.toPath(), tempFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
-      } catch (IOException e) {
-        try {
-          tempFile.delete();
-        } catch (Exception e0) {
-          logger.debug("delete temp file error, e: ", e0);
-        }
-        logger.error(
-            "copy dynamic lib from jar failed, file: {}, e: ", tempFile.getAbsolutePath(), e);
-        throw e;
-      } catch (NullPointerException e) {
-        throw new FileNotFoundException("Cannot found " + libName + " inside the JAR.");
+    try (InputStream is =
+        JniLibLoader.class.getResourceAsStream(NATIVE_WIN_DEPS_DIR + WIN_DEPS_FILE_LIST)) {
+      if (is == null) {
+        throw new FileNotFoundException("Cannot found " + WIN_DEPS_FILE_LIST + " inside the JAR.");
       }
-      try {
-        loadLibrary(tempFile.getAbsolutePath(), true);
-      } catch (Exception e) {
-        logger.error("loadLibrary error, resource: {}, e: ", tempFile.getAbsolutePath(), e);
-        throw e;
-      } finally {
-        try {
-          tempFile.delete();
-          tempDir.delete();
-          logger.debug("remove temp dir and temp file, {}", tempFile);
-        } catch (Exception ignored) {
+      try (Scanner scanner = new Scanner(is)) {
+        while (scanner.hasNext()) {
+          String lib = scanner.next();
+          File targetFile = new File(tempDir, lib);
+          logger.info("loadWinDepsLibraryFromJar tempDir: {}, targetFile: {}", tempDir, targetFile);
+          try (InputStream fis =
+              JniLibLoader.class.getResourceAsStream(NATIVE_WIN_DEPS_DIR + lib)) {
+            if (fis == null) {
+              throw new FileNotFoundException("Cannot found " + lib + " inside the JAR.");
+            }
+            Files.copy(fis, targetFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+            loadLibrary(targetFile.getAbsolutePath(), true);
+          } catch (IOException e) {
+            targetFile.delete();
+          } catch (Exception e) {
+            logger.error("loadLibrary error, resource: {}, e: ", targetFile.getAbsolutePath(), e);
+            throw e;
+          } finally {
+            try {
+              targetFile.delete();
+              tempDir.delete();
+              logger.debug("remove temp dir and temp file, {}", targetFile);
+            } catch (Exception ignored) {
+            }
+          }
         }
       }
     }
